replacing size_t with int

+++ source/data-structures/bit.h
@@ -1,20 +1,20 @@
 
 template<typename T>
 struct bit {
-    size_t n, lg;
+    int n, lg;
     vector<T> a;
-    bit(size_t n) : n(n) {
+    bit(int n) : n(n) {
         a.assign(n, T());
         lg = 0;
         while(2 * (1u << lg) < n) lg++;
     }
-    void add(size_t i, T x) {
+    void add(int i, T x) {
         while(i < n) {
             a[i] = a[i] + x;
             i += i & -i;
         }
     }
-    T get(size_t i) {
+    T get(int i) {
         T x = T();
         while(i > 0) {
             x = x + a[i];
@@ -22,10 +22,10 @@ struct bit {
         }
         return x;
     }
-    size_t lower_bound(T val) {
-        size_t i = 0;
+    int lower_bound(T val) {
+        int i = 0;
         T sum = T();
-        for(size_t k = lg; ; k--) {
+        for(int k = lg; ; k--) {
             if(i + (1 << k) < n && sum + a[i + (1 << k)] < val) {
                 i += (1 << k);
                 sum = sum + a[i];
